backend/requirements.txt
Django==5.1.5
numpy==1.26.4
djangorestframework == 3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
ruff==0.6.7
django-filter==24.3
django-cors-headers==4.4.0
django-silk==5.2.0
django-allauth==65.0.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
oauthlib==3.2.2
requests-oauthlib==2.0.0
openai==1.97.0
drf-extra-fields==3.7.0
psycopg2-binary==2.9.9
opencv-python==4.9.0.80
